requirements.txt
absl-py==0.13.0
albumentations==1.0.3
argon2-cffi==20.1.0
asttokens==2.0.5
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.0.0
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.14.6
charset-normalizer==2.0.4
chumpy==0.70
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.10.0
cytoolz==0.11.0
dask==2021.10.0
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
easydict==1.9
einops==0.3.2
entrypoints==0.3
executing==0.8.3
face-alignment==1.3.4
filelock==3.3.2
fonttools==4.25.0
fsspec==2021.8.1
future==0.18.2
fvcore==0.1.5.post20210915
gdown==4.2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
icecream==2.1.2
idna==3.2
imagecodecs==2021.8.26
imageio==2.9.0
imageio-ffmpeg==0.4.5
imgaug==0.4.0
importlib-metadata==4.8.1
iopath==0.1.9
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
kmeans-pytorch==0.3
kornia==0.6.1
llvmlite==0.37.0
lmdb==1.2.1
locket==0.2.1
lpips==0.1.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.22.0
oauthlib==3.1.1
olefile==0.46
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
packaging==21.0
pandas==1.1.3
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
portalocker==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.20
Pygments==2.10.0
pyhocon==0.3.59
PyMCubes==0.1.2
pyntcloud==0.2.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pythreejs==2.3.0
pytorch-msssim==0.2.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.2
Send2Trash==1.8.0
setuptools==58.0.4
Shapely==1.7.1
six==1.16.0
sklearn==0.0
soupsieve==2.3
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
toolz==0.11.1
torchsummary==1.5.1
tqdm==4.62.3
traitlets==5.1.0
traittypes==0.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
yacs==0.1.6
zipp==3.6.0
